20190813 RelatedQuestions

+++ src/main/resources/db/migration/V2__Create_question_table.sql
@@ -1,11 +1,11 @@
 create table question
 (
-	id int auto_increment,
+	id BIGINT auto_increment,
 	title varchar(50),
 	description TEXT,
 	gmt_create BIGINT,
 	gmt_modified BIGINT,
-	creator int,
+	creator BIGINT,
 	comment_count int default 0,
 	view_count int default 0,
 	like_count int default 0,
